# Speed up TM simulation