--- a/src/tm/tm2/tm2_meta/tm2_simulator.py
+++ b/src/tm/tm2/tm2_meta/tm2_simulator.py
@@ -26,7 +26,7 @@ if __name__ == "__main__":
 
     quiet = ("-q" in args)
 
-    numSteps = float("Inf") # default value
+    numSteps = sys.maxint # default value
     if ("-s" in args):
         numSteps = args[args.index("-s") + 1]