--- a/src/tm/tm2/tm2_meta/tmsim.py
+++ b/src/tm/tm2/tm2_meta/tmsim.py
@@ -97,7 +97,7 @@ def stateDictionariesToLists(stateDictionary, alphabet, startState):
             simulationStates[state]._initFromState(state, simulationStates)
     return simulationStates[startState]
 
-class SingleTapeTuringMachine:
+class SingleTapeTuringMachine(object):
     def __init__(self, path, alphabet=["_", "1", "H", "E"]):        
         self.state = None
         self.tape = Tape(None, alphabet[0])