Hi Adam,

congrats on your work ;-). I guess that you ultimately don't care about the speed of Turing machine performance too much, since the point isn't to run them anyway. But I felt like it, so I optimized it somewhat. This is a bit at the expense of readability, so feel free to just reject the pull request.

As performance measurements I used 10**8 steps of ``goldbach`` and running ``squaresaresmall`` till it halts, with the 2 symbol version (the 4 symbol version has very similar results). I ran these on PyPy and Goldbach on CPython (not enough patience for ``squaresaresmall``). The results improved as follows:

VM | Benchmark | old (s) | new (s) | factor
---|-----------|---------|---------|-------
PyPy | Goldbach | 30.8 | 1.4 | 22x
CPython | Goldbach | 300.9 | 137.8 | 2.2x
PyPy | squaresaresmall | 363.0 | 31.9 | 11.4x

This makes the code on PyPy a 70 MHz Turing machine emulator on my laptop ;-).

Cheers,

Carl Friedrich