@arigo 2016-05-13
Inspired me to try a simple .tm2-to-.c translator, using gcc to compile the final programs.  It runs squaresaresmall in 1.3 seconds.  Program attached to this issue (feel free to use or to ignore): run "python parse.py squaresaresmall.tm2 > tm.i && gcc -O2 tm.c".

[tm2-to-c.zip](https://github.com/adamyedidia/parsimony/files/263931/tm2-to-c.zip)