@adamyedidia 2016-05-13
Whoa, impressive! Thanks for doing this work. When you say that it's at the expense of readability, are you referring to the readability of the code, or did you change the output format at all?