@cfbolz 2016-05-13
no, I mean the readability of ``tmsim.py``. Input output should be the same.