@adamyedidia 2016-05-13
Thanks very much to both of you! I have added both your contributions to the project. :)